+++ freight/trips.py
@@ -4,14 +4,35 @@ import numpy as np
 def configure(context, require):
     require.stage("freight.gte.trips")
     require.stage("freight.gqgv.trips")
+    require.stage("data.freight.departure_times")
 
 def execute(context):
     df_gte_trips = context.stage("freight.gte.trips")
     df_gqgv_trips = context.stage("freight.gqgv.trips")
+    df_departure_times = context.stage("data.freight.departure_times")
 
     df_trips = pd.concat([df_gte_trips, df_gqgv_trips], ignore_index=True)
 
-    # reset agent ids
+    # generate hour from departure time distribution
+    departure_times = 3600 * np.random.choice(df_departure_times["hour"].values,
+                                       len(df_trips),
+                                       p=df_departure_times["probability"].values)
+
+    # add seconds with hour randomly from uniform distribution
+    departure_times += (3600 * np.random.rand(len(df_trips))).astype(int)
+
+    # add departure times to trips
+    df_trips["departure_time"] = departure_times
+
+    # add agent ids
     df_trips["agent_id"] = np.arange(len(df_trips))
+
+    # package up
+    df_trips = df_trips[["agent_id",
+                         "origin_x", "origin_y",
+                         "destination_x", "destination_y",
+                         "departure_time",
+                         "vehicle_type"]
+    ]
 
     return df_trips